Rename Inode->MutableInode, ReadOnlyInode->Inode
