+++ core/server/master/src/main/java/alluxio/master/file/DefaultPermissionChecker.java
@@ -17,8 +17,8 @@ import alluxio.exception.AccessControlException;
 import alluxio.exception.ExceptionMessage;
 import alluxio.exception.InvalidPathException;
 import alluxio.exception.PreconditionMessage;
+import alluxio.master.file.meta.Inode;
 import alluxio.master.file.meta.InodeTree;
-import alluxio.master.file.meta.ReadOnlyInode;
 import alluxio.master.file.meta.LockedInodePath;
 import alluxio.security.authentication.AuthenticatedClientUser;
 import alluxio.security.authorization.AclAction;
@@ -74,7 +74,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
 
     // collects existing inodes info on the path. Note that, not all the components of the path have
     // corresponding inodes.
-    List<ReadOnlyInode> inodeList = inodePath.getInodeList();
+    List<Inode> inodeList = inodePath.getInodeList();
 
     // collects user and groups
     String user = AuthenticatedClientUser.getClientUser();
@@ -95,7 +95,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
     }
 
     // collects inodes info on the path
-    List<ReadOnlyInode> inodeList = inodePath.getInodeList();
+    List<Inode> inodeList = inodePath.getInodeList();
 
     // collects user and groups
     String user = AuthenticatedClientUser.getClientUser();
@@ -110,7 +110,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
       return Mode.Bits.NONE;
     }
     // collects inodes info on the path
-    List<ReadOnlyInode> inodeList = inodePath.getInodeList();
+    List<Inode> inodeList = inodePath.getInodeList();
 
     // collects user and groups
     try {
@@ -164,7 +164,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
   private void checkOwner(LockedInodePath inodePath)
       throws AccessControlException, InvalidPathException {
     // collects inodes info on the path
-    List<ReadOnlyInode> inodeList = inodePath.getInodeList();
+    List<Inode> inodeList = inodePath.getInodeList();
 
     // collects user and groups
     String user = AuthenticatedClientUser.getClientUser();
@@ -206,9 +206,8 @@ public class DefaultPermissionChecker implements PermissionChecker {
    * @param checkIsOwner indicates whether to check the user is the owner of the path
    * @throws AccessControlException if permission checking fails
    */
-  protected void checkInodeList(String user, List<String> groups, Mode.Bits bits,
-      String path, List<ReadOnlyInode> inodeList, boolean checkIsOwner)
-      throws AccessControlException {
+  protected void checkInodeList(String user, List<String> groups, Mode.Bits bits, String path,
+      List<Inode> inodeList, boolean checkIsOwner) throws AccessControlException {
     int size = inodeList.size();
     Preconditions.checkArgument(size > 0,
         PreconditionMessage.EMPTY_FILE_INFO_LIST_FOR_PERMISSION_CHECK);
@@ -223,7 +222,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
       checkInode(user, groups, inodeList.get(i), Mode.Bits.EXECUTE, path);
     }
 
-    ReadOnlyInode inode = inodeList.get(inodeList.size() - 1);
+    Inode inode = inodeList.get(inodeList.size() - 1);
     if (checkIsOwner) {
       if (inode == null || user.equals(inode.getOwner())) {
         return;
@@ -244,8 +243,8 @@ public class DefaultPermissionChecker implements PermissionChecker {
    * @param path the path to check permission on
    * @throws AccessControlException if permission checking fails
    */
-  private void checkInode(String user, List<String> groups, ReadOnlyInode inode, Mode.Bits bits,
-      String path) throws AccessControlException {
+  private void checkInode(String user, List<String> groups, Inode inode, Mode.Bits bits,
+           String path) throws AccessControlException {
     if (inode == null) {
       return;
     }
@@ -267,7 +266,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
    * @return the permission
    */
   private Mode.Bits getPermissionInternal(String user, List<String> groups, String path,
-      List<ReadOnlyInode> inodeList) {
+      List<Inode> inodeList) {
     int size = inodeList.size();
     Preconditions.checkArgument(size > 0,
         PreconditionMessage.EMPTY_FILE_INFO_LIST_FOR_PERMISSION_CHECK);
@@ -286,7 +285,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
       }
     }
 
-    ReadOnlyInode inode = inodeList.get(inodeList.size() - 1);
+    Inode inode = inodeList.get(inodeList.size() - 1);
     if (inode == null) {
       return Mode.Bits.NONE;
     }
@@ -299,7 +298,7 @@ public class DefaultPermissionChecker implements PermissionChecker {
   }
 
   private static String toExceptionMessage(String user, Mode.Bits bits, String path,
-      ReadOnlyInode inode) {
+      Inode inode) {
     StringBuilder sb =
         new StringBuilder().append("user=").append(user).append(", ").append("access=").append(bits)
             .append(", ").append("path=").append(path).append(": ").append("failed at ")
