+++ core/server/master/src/main/java/alluxio/master/file/meta/Inode.java
@@ -11,444 +11,165 @@
 
 package alluxio.master.file.meta;
 
-import alluxio.Constants;
-import alluxio.master.ProtobufUtils;
-import alluxio.proto.journal.File.UpdateInodeEntry;
+import alluxio.proto.journal.Journal.JournalEntry;
 import alluxio.proto.meta.InodeMeta;
-import alluxio.proto.meta.InodeMeta.InodeOrBuilder;
 import alluxio.security.authorization.AccessControlList;
 import alluxio.security.authorization.AclAction;
 import alluxio.security.authorization.AclActions;
-import alluxio.security.authorization.AclEntry;
-import alluxio.security.authorization.AclEntryType;
 import alluxio.security.authorization.DefaultAccessControlList;
 import alluxio.wire.FileInfo;
 import alluxio.wire.TtlAction;
 
-import com.google.common.base.Objects;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import javax.annotation.concurrent.NotThreadSafe;
-
 /**
- * {@link Inode} is an abstract class, with information shared by all types of Inodes.
- *
- * @param <T> the concrete subclass of this object
+ * Base class for read only inodes.
  */
-@NotThreadSafe
-public abstract class Inode<T extends Inode> implements InodeView {
-  private static final Logger LOG = LoggerFactory.getLogger(Inode.class);
-  protected long mCreationTimeMs;
-  private boolean mDeleted;
-  protected final boolean mDirectory;
-  protected final long mId;
-  protected long mTtl;
-  protected TtlAction mTtlAction;
-  private long mLastModificationTimeMs;
-  private String mName;
-  private long mParentId;
-  private PersistenceState mPersistenceState;
-  private boolean mPinned;
-  protected AccessControlList mAcl;
-  private String mUfsFingerprint;
+public abstract class Inode implements InodeView {
+  private final InodeView mDelegate;
 
-  protected Inode(long id, boolean isDirectory) {
-    mCreationTimeMs = System.currentTimeMillis();
-    mDeleted = false;
-    mDirectory = isDirectory;
-    mId = id;
-    mTtl = Constants.NO_TTL;
-    mTtlAction = TtlAction.DELETE;
-    mLastModificationTimeMs = mCreationTimeMs;
-    mName = null;
-    mParentId = InodeTree.NO_PARENT;
-    mPersistenceState = PersistenceState.NOT_PERSISTED;
-    mPinned = false;
-    mAcl = new AccessControlList();
-    mUfsFingerprint = Constants.INVALID_UFS_FINGERPRINT;
+  protected Inode(InodeView delegate) {
+    mDelegate = delegate;
   }
 
   @Override
   public long getCreationTimeMs() {
-    return mCreationTimeMs;
+    return mDelegate.getCreationTimeMs();
   }
 
   @Override
   public String getGroup() {
-    return mAcl.getOwningGroup();
+    return mDelegate.getGroup();
   }
 
   @Override
   public long getId() {
-    return mId;
+    return mDelegate.getId();
   }
 
   @Override
   public long getTtl() {
-    return mTtl;
+    return mDelegate.getTtl();
   }
 
   @Override
   public TtlAction getTtlAction() {
-    return mTtlAction;
+    return mDelegate.getTtlAction();
   }
 
   @Override
   public long getLastModificationTimeMs() {
-    return mLastModificationTimeMs;
+    return mDelegate.getLastModificationTimeMs();
   }
 
   @Override
   public String getName() {
-    return mName;
+    return mDelegate.getName();
   }
 
   @Override
   public short getMode() {
-    return mAcl.getMode();
+    return mDelegate.getMode();
   }
 
   @Override
   public PersistenceState getPersistenceState() {
-    return mPersistenceState;
+    return mDelegate.getPersistenceState();
   }
 
   @Override
   public long getParentId() {
-    return mParentId;
+    return mDelegate.getParentId();
   }
 
   @Override
   public String getOwner() {
-    return mAcl.getOwningUser();
+    return mDelegate.getOwner();
   }
 
   @Override
   public boolean isDeleted() {
-    return mDeleted;
+    return mDelegate.isDeleted();
   }
 
   @Override
   public boolean isDirectory() {
-    return mDirectory;
+    return mDelegate.isDirectory();
   }
 
   @Override
   public boolean isFile() {
-    return !mDirectory;
+    return mDelegate.isFile();
   }
 
   @Override
   public boolean isPinned() {
-    return mPinned;
+    return mDelegate.isPinned();
   }
 
   @Override
   public boolean isPersisted() {
-    return mPersistenceState == PersistenceState.PERSISTED;
+    return mDelegate.isPersisted();
   }
 
   @Override
   public String getUfsFingerprint() {
-    return mUfsFingerprint;
+    return mDelegate.getUfsFingerprint();
   }
 
   @Override
   public AccessControlList getACL() {
-    return mAcl;
-  }
-
-  /**
-   * Removes the extended ACL entries. The base entries are retained.
-   *
-   * @return the updated object
-   */
-  public T removeExtendedAcl() {
-    mAcl.removeExtendedEntries();
-    return getThis();
-  }
-
-  /**
-   * Removes ACL entries.
-   *
-   * @param entries the ACL entries to remove
-   * @return the updated object
-   */
-  public T removeAcl(List<AclEntry> entries) {
-    for (AclEntry entry : entries) {
-      if (entry.isDefault()) {
-        AccessControlList defaultAcl = getDefaultACL();
-        defaultAcl.removeEntry(entry);
-      } else {
-        mAcl.removeEntry(entry);
-      }
-    }
-    updateMask(entries);
-    return getThis();
-  }
-
-  /**
-   * Replaces all existing ACL entries with a new list of entries.
-   *
-   * @param entries the new list of ACL entries
-   * @return the updated object
-   */
-  public T replaceAcl(List<AclEntry> entries) {
-    boolean clearACL = false;
-    for (AclEntry entry : entries) {
-      /**
-       * if we are only setting default ACLs, we do not need to clear access ACL entries
-       * observed same behavior on linux
-       */
-      if (!entry.isDefault()) {
-        clearACL = true;
-      }
-    }
-    if (clearACL) {
-      mAcl.clearEntries();
-    }
-    return setAcl(entries);
-  }
-
-  /**
-   * Update Mask for the Inode.
-   * This method should be called after updates to ACL and defaultACL.
-   *
-   * @param entries the list of ACL entries
-   * @return the updated object
-   */
-  public T updateMask(List<AclEntry> entries) {
-    boolean needToUpdateACL = false;
-    boolean needToUpdateDefaultACL = false;
-
-    for (AclEntry entry : entries) {
-      if (entry.getType().equals(AclEntryType.NAMED_USER)
-          || entry.getType().equals(AclEntryType.NAMED_GROUP)
-          || entry.getType().equals(AclEntryType.OWNING_GROUP)) {
-        if (entry.isDefault()) {
-          needToUpdateDefaultACL = true;
-        } else {
-          needToUpdateACL = true;
-        }
-      }
-      if (entry.getType().equals(AclEntryType.MASK)) {
-        // If mask is explicitly set or removed then we don't need to update the mask
-        return getThis();
-      }
-    }
-    if (needToUpdateACL) {
-      mAcl.updateMask();
-    }
-
-    if (needToUpdateDefaultACL) {
-      getDefaultACL().updateMask();
-    }
-    return getThis();
-  }
-
-  /**
-   * @param creationTimeMs the creation time to use (in milliseconds)
-   * @return the updated object
-   */
-  public T setCreationTimeMs(long creationTimeMs) {
-    mCreationTimeMs = creationTimeMs;
-    return getThis();
-  }
-
-  /**
-   * @param deleted the deleted flag to use
-   * @return the updated object
-   */
-  public T setDeleted(boolean deleted) {
-    mDeleted = deleted;
-    return getThis();
-  }
-
-  /**
-   * @param group the group of the inode
-   * @return the updated object
-   */
-  public T setGroup(String group) {
-    mAcl.setOwningGroup(group);
-    if (isDirectory()) {
-      getDefaultACL().setOwningGroup(group);
-    }
-    return getThis();
-  }
-
-  /**
-   * Sets the last modification time of the inode to the new time if the new time is more recent.
-   * This method can be called concurrently with deterministic results.
-   *
-   * @param lastModificationTimeMs the last modification time to use
-   * @return the updated object
-   */
-  public T setLastModificationTimeMs(long lastModificationTimeMs) {
-    return setLastModificationTimeMs(lastModificationTimeMs, false);
-  }
-
-  /**
-   * @param lastModificationTimeMs the last modification time to use
-   * @param override if true, sets the value regardless of the previous last modified time,
-   *                 should be set to true for journal replay
-   * @return the updated object
-   */
-  public T setLastModificationTimeMs(long lastModificationTimeMs, boolean override) {
-    synchronized (this) {
-      if (override || mLastModificationTimeMs < lastModificationTimeMs) {
-        mLastModificationTimeMs = lastModificationTimeMs;
-      }
-      return getThis();
-    }
-  }
-
-  /**
-   * @param name the name to use
-   * @return the updated object
-   */
-  public T setName(String name) {
-    mName = name;
-    return getThis();
-  }
-
-  /**
-   * @param parentId the parent id to use
-   * @return the updated object
-   */
-  public T setParentId(long parentId) {
-    mParentId = parentId;
-    return getThis();
-  }
-
-  /**
-   * @param ttl the TTL to use, in milliseconds
-   * @return the updated object
-   */
-  public T setTtl(long ttl) {
-    mTtl = ttl;
-    return getThis();
-  }
-
-  /**
-   * @param ttlAction the {@link TtlAction} to use
-   * @return the updated options object
-   */
-  public T setTtlAction(TtlAction ttlAction) {
-    mTtlAction = ttlAction;
-    return getThis();
-  }
-
-  /**
-   * @param persistenceState the {@link PersistenceState} to use
-   * @return the updated object
-   */
-  public T setPersistenceState(PersistenceState persistenceState) {
-    mPersistenceState = persistenceState;
-    return getThis();
-  }
-
-  /**
-   * @param pinned the pinned flag value to use
-   * @return the updated object
-   */
-  public T setPinned(boolean pinned) {
-    mPinned = pinned;
-    return getThis();
-  }
-
-  /**
-   * @param owner the owner name of the inode
-   * @return the updated object
-   */
-  public T setOwner(String owner) {
-    mAcl.setOwningUser(owner);
-    if (isDirectory()) {
-      getDefaultACL().setOwningUser(owner);
-    }
-    return getThis();
-  }
-
-  /**
-   * @param mode the mode of the inode
-   * @return the updated object
-   */
-  public T setMode(short mode) {
-    mAcl.setMode(mode);
-    return getThis();
-  }
-
-  /**
-   * @param acl set the default ACL associated with this inode
-   * @throws UnsupportedOperationException if the inode is a file
-   * @return the updated object
-   */
-  public abstract T setDefaultACL(DefaultAccessControlList acl)
-      throws UnsupportedOperationException;
-
-  /**
-   * Sets ACL entries into the internal ACL.
-   * The entries will overwrite any existing correspondent entries in the internal ACL.
-   *
-   * @param entries the ACL entries
-   * @return the updated object
-   */
-  public T setAcl(List<AclEntry> entries) {
-    if (entries == null || entries.isEmpty()) {
-      return getThis();
-    }
-    for (AclEntry entry : entries) {
-      if (entry.isDefault()) {
-        getDefaultACL().setEntry(entry);
-      } else {
-        mAcl.setEntry(entry);
-      }
-    }
-    updateMask(entries);
-    return getThis();
-  }
-
-  /**
-   * Sets the internal ACL to a specified ACL.
-   * @param acl the specified ACL
-   * @return the updated object
-   */
-  public T setInternalAcl(AccessControlList acl) {
-    mAcl = acl;
-    return getThis();
-  }
-
-  /**
-   * @param ufsFingerprint the ufs fingerprint to use
-   * @return the updated object
-   */
-  public T setUfsFingerprint(String ufsFingerprint) {
-    mUfsFingerprint = ufsFingerprint;
-    return getThis();
+    return mDelegate.getACL();
   }
 
   @Override
-  public abstract FileInfo generateClientFileInfo(String path);
+  public DefaultAccessControlList getDefaultACL() throws UnsupportedOperationException {
+    return mDelegate.getDefaultACL();
+  }
 
-  /**
-   * @return {@code this} so that the abstract class can use the fluent builder pattern
-   */
-  protected abstract T getThis();
+  @Override
+  public FileInfo generateClientFileInfo(String path) {
+    return mDelegate.generateClientFileInfo(path);
+  }
 
   @Override
   public boolean checkPermission(String user, List<String> groups, AclAction action) {
-    return mAcl.checkPermission(user, groups, action);
+    return mDelegate.checkPermission(user, groups, action);
   }
 
   @Override
   public AclActions getPermission(String user, List<String> groups) {
-    return mAcl.getPermission(user, groups);
+    return mDelegate.getPermission(user, groups);
+  }
+
+  @Override
+  public InodeMeta.Inode toProto() {
+    return mDelegate.toProto();
+  }
+
+  @Override
+  public JournalEntry toJournalEntry() {
+    return mDelegate.toJournalEntry();
+  }
+
+  @Override
+  public boolean equals(Object other) {
+    if (!(other instanceof Inode)) {
+      return false;
+    }
+    Inode otherInode = (Inode) other;
+    return mDelegate.equals(otherInode.mDelegate);
+  }
+
+  @Override
+  public int hashCode() {
+    return mDelegate.hashCode();
   }
 
   /**
-   * Casts the inode as an {@link ReadOnlyInodeDirectory} if it is one, otherwise throws an
+   * Casts the inode as an {@link InodeDirectory} if it is one, otherwise throws an
    * exception.
    *
    * This gives convenience in method chaining, e.g.
@@ -463,7 +184,8 @@ public abstract class Inode<T extends Inode> implements InodeView {
    */
   public InodeDirectory asDirectory() {
     if (!isDirectory()) {
-      throw new IllegalStateException(String.format("Inode %s is not a directory", getName()));
+      throw new IllegalStateException(
+          String.format("Inode %s is not a directory", mDelegate.getName()));
     }
     return (InodeDirectory) this;
   }
@@ -473,113 +195,29 @@ public abstract class Inode<T extends Inode> implements InodeView {
    */
   public InodeFile asFile() {
     if (isDirectory()) {
-      throw new IllegalStateException(String.format("Inode %s is not a file", getName()));
+      throw new IllegalStateException(
+          String.format("Inode %s is not a file", mDelegate.getName()));
     }
     return (InodeFile) this;
   }
 
   /**
-   * Updates this inode's state from the given entry.
+   * Wraps an InodeView, providing read-only access. Modifications to the underlying inode will
+   * affect the created read-only inode.
    *
-   * @param entry the entry
+   * @param delegate the delegate to wrap
+   * @return the created read-only inode
    */
-  public void updateFromEntry(UpdateInodeEntry entry) {
-    if (entry.hasAcl()) {
-      setInternalAcl(AccessControlList.fromProtoBuf(entry.getAcl()));
+  public static Inode wrap(InodeView delegate) {
+    if (delegate instanceof Inode) {
+      return (Inode) delegate;
     }
-    if (entry.hasCreationTimeMs()) {
-      setCreationTimeMs(entry.getCreationTimeMs());
-    }
-    if (entry.hasGroup()) {
-      setGroup(entry.getGroup());
-    }
-    if (entry.hasLastModificationTimeMs()) {
-      setLastModificationTimeMs(entry.getLastModificationTimeMs(),
-          entry.getOverwriteModificationTime());
-    }
-    if (entry.hasMode()) {
-      setMode((short) entry.getMode());
-    }
-    if (entry.hasName()) {
-      setName(entry.getName());
-    }
-    if (entry.hasOwner()) {
-      setOwner(entry.getOwner());
-    }
-    if (entry.hasParentId()) {
-      setParentId(entry.getParentId());
-    }
-    if (entry.hasPersistenceState()) {
-      setPersistenceState(PersistenceState.valueOf(entry.getPersistenceState()));
-    }
-    if (entry.hasPinned()) {
-      setPinned(entry.getPinned());
-    }
-    if (entry.hasTtl()) {
-      setTtl(entry.getTtl());
-    }
-    if (entry.hasTtlAction()) {
-      setTtlAction(ProtobufUtils.fromProtobuf(entry.getTtlAction()));
-    }
-    if (entry.hasUfsFingerprint()) {
-      setUfsFingerprint(entry.getUfsFingerprint());
-    }
-  }
-
-  @Override
-  public int hashCode() {
-    return ((Long) mId).hashCode();
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (!(o instanceof Inode<?>)) {
-      return false;
-    }
-    Inode<?> that = (Inode<?>) o;
-    return mId == that.mId;
-  }
-
-  protected Objects.ToStringHelper toStringHelper() {
-    return Objects.toStringHelper(this).add("id", mId).add("name", mName).add("parentId", mParentId)
-        .add("creationTimeMs", mCreationTimeMs).add("pinned", mPinned).add("deleted", mDeleted)
-        .add("ttl", mTtl).add("ttlAction", mTtlAction)
-        .add("directory", mDirectory).add("persistenceState", mPersistenceState)
-        .add("lastModificationTimeMs", mLastModificationTimeMs)
-        .add("owner", mAcl.getOwningUser())
-        .add("group", mAcl.getOwningGroup())
-        .add("permission", mAcl.getMode())
-        .add("ufsFingerprint", mUfsFingerprint);
-  }
-
-  protected InodeMeta.Inode.Builder toProtoBuilder() {
-    return InodeMeta.Inode.newBuilder()
-        .setId(getId())
-        .setCreationTimeMs(getCreationTimeMs())
-        .setIsDirectory(isDirectory())
-        .setTtl(getTtl())
-        .setTtlAction(ProtobufUtils.toProtobuf(getTtlAction()))
-        .setLastModifiedMs(getLastModificationTimeMs())
-        .setName(getName())
-        .setParentId(getParentId())
-        .setPersistenceState(getPersistenceState().name())
-        .setIsPinned(isPinned())
-        .setAccessAcl(AccessControlList.toProtoBuf(getACL()))
-        .setUfsFingerprint(getUfsFingerprint());
-  }
-
-  /**
-   * @param inode a protocol buffer inode
-   * @return the {@link Inode} representation for the inode
-   */
-  public static Inode<?> fromProto(InodeOrBuilder inode) {
-    if (inode.getIsDirectory()) {
-      return InodeDirectory.fromProto(inode);
+    if (delegate.isFile()) {
+      Preconditions.checkState(delegate instanceof InodeFileView);
+      return new InodeFile((InodeFileView) delegate);
     } else {
-      return InodeFile.fromProto(inode);
+      Preconditions.checkState(delegate instanceof InodeDirectoryView);
+      return new InodeDirectory((InodeDirectoryView) delegate);
     }
   }
 }
