+++ core/server/master/src/main/java/alluxio/master/file/meta/TtlBucketList.java
@@ -70,14 +70,14 @@ public final class TtlBucketList {
   }
 
   /**
-   * Inserts an {@link Inode} to the appropriate bucket where its ttl end time lies in the
+   * Inserts an inode to the appropriate bucket where its ttl end time lies in the
    * bucket's interval, if no appropriate bucket exists, a new bucket will be created to contain
    * this inode, if ttl value is {@link Constants#NO_TTL}, the inode won't be inserted to any
    * buckets and nothing will happen.
    *
    * @param inode the inode to be inserted
    */
-  public void insert(ReadOnlyInode inode) {
+  public void insert(Inode inode) {
     if (inode.getTtl() == Constants.NO_TTL) {
       return;
     }
