+++ core/server/master/src/main/java/alluxio/master/file/meta/UfsSyncUtils.java
@@ -12,7 +12,6 @@
 package alluxio.master.file.meta;
 
 import alluxio.underfs.Fingerprint;
-import alluxio.underfs.UfsStatus;
 
 import javax.annotation.concurrent.NotThreadSafe;
 
@@ -25,7 +24,7 @@ public final class UfsSyncUtils {
   private UfsSyncUtils() {} // prevent instantiation
 
   /**
-   * Given an {@link Inode} and {@link UfsStatus}, returns a {@link SyncPlan} describing how to
+   * Given an inode and ufs status, returns a sync plan describing how to
    * sync the inode with the ufs.
    *
    * @param inode the inode to sync
@@ -33,7 +32,7 @@ public final class UfsSyncUtils {
    * @param containsMountPoint true if this inode contains a mount point, false otherwise
    * @return a {@link SyncPlan} describing how to sync the inode with the ufs
    */
-  public static SyncPlan computeSyncPlan(ReadOnlyInode inode, Fingerprint ufsFingerprint,
+  public static SyncPlan computeSyncPlan(Inode inode, Fingerprint ufsFingerprint,
       boolean containsMountPoint) {
     Fingerprint inodeFingerprint =  Fingerprint.parse(inode.getUfsFingerprint());
     boolean isContentSynced = inodeUfsIsContentSynced(inode, inodeFingerprint, ufsFingerprint);
@@ -93,7 +92,7 @@ public final class UfsSyncUtils {
    * @param ufsFingerprint the ufs fingerprint to check for the sync
    * @return true of the inode is synced with the ufs status
    */
-  public static boolean inodeUfsIsContentSynced(ReadOnlyInode inode, Fingerprint inodeFingerprint,
+  public static boolean inodeUfsIsContentSynced(Inode inode, Fingerprint inodeFingerprint,
       Fingerprint ufsFingerprint) {
     boolean isSyncedUnpersisted =
         !inode.isPersisted() && !ufsFingerprint.isValid();
@@ -114,8 +113,8 @@ public final class UfsSyncUtils {
    * @param ufsFingerprint the ufs fingerprint to check for the sync
    * @return true of the inode is synced with the ufs status
    */
-  public static boolean inodeUfsIsMetadataSynced(ReadOnlyInode inode, Fingerprint inodeFingerprint,
-      Fingerprint ufsFingerprint) {
+  public static boolean inodeUfsIsMetadataSynced(Inode inode, Fingerprint inodeFingerprint,
+         Fingerprint ufsFingerprint) {
     return inodeFingerprint.isValid() && inodeFingerprint.matchMetadata(ufsFingerprint);
   }
