+++ core/server/master/src/test/java/alluxio/master/file/meta/BaseInodeLockingTest.java
@@ -37,12 +37,12 @@ public class BaseInodeLockingTest {
   protected InodeStore mInodeStore = new HeapInodeStore();
 
   // Directory structure is /a/b/c
-  protected ReadOnlyInodeFile mFileC = inodeFile(3, 2, "c");
-  protected ReadOnlyInodeDirectory mDirB = inodeDir(2, 1, "b", mFileC);
-  protected ReadOnlyInodeDirectory mDirA = inodeDir(1, 0, "a", mDirB);
-  protected ReadOnlyInodeDirectory mRootDir = inodeDir(0, -1, "", mDirA);
+  protected InodeFile mFileC = inodeFile(3, 2, "c");
+  protected InodeDirectory mDirB = inodeDir(2, 1, "b", mFileC);
+  protected InodeDirectory mDirA = inodeDir(1, 0, "a", mDirB);
+  protected InodeDirectory mRootDir = inodeDir(0, -1, "", mDirA);
 
-  protected List<ReadOnlyInode> mAllInodes = Arrays.asList(mRootDir, mDirA, mDirB, mFileC);
+  protected List<Inode> mAllInodes = Arrays.asList(mRootDir, mDirA, mDirB, mFileC);
 
   @After
   public void after() {
@@ -56,13 +56,13 @@ public class BaseInodeLockingTest {
   /**
    * Checks that only the specified inodes are read-locked.
    */
-  protected void checkOnlyNodesReadLocked(ReadOnlyInode... inodes) {
-    HashSet<ReadOnlyInode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
-    for (ReadOnlyInode inode : inodes) {
+  protected void checkOnlyNodesReadLocked(Inode... inodes) {
+    HashSet<Inode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
+    for (Inode inode : inodes) {
       assertTrue("Expected inode " + inode.getId() + " to be read locked",
           mInodeLockManager.inodeReadLockedByCurrentThread(inode.getId()));
     }
-    for (ReadOnlyInode inode : mAllInodes) {
+    for (Inode inode : mAllInodes) {
       if (!shouldBeLocked.contains(inode)) {
         assertFalse("Expected inode " + inode.getId() + " to not be read locked",
             mInodeLockManager.inodeReadLockedByCurrentThread(inode.getId()));
@@ -73,13 +73,13 @@ public class BaseInodeLockingTest {
   /**
    * Checks that only the specified inodes are write-locked.
    */
-  protected void checkOnlyNodesWriteLocked(ReadOnlyInode... inodes) {
-    HashSet<ReadOnlyInode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
-    for (ReadOnlyInode inode : inodes) {
+  protected void checkOnlyNodesWriteLocked(Inode... inodes) {
+    HashSet<Inode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
+    for (Inode inode : inodes) {
       assertTrue("Expected inode " + inode.getId() + " to be write locked",
           mInodeLockManager.inodeWriteLockedByCurrentThread(inode.getId()));
     }
-    for (ReadOnlyInode inode : mAllInodes) {
+    for (Inode inode : mAllInodes) {
       if (!shouldBeLocked.contains(inode)) {
         assertFalse("Expected inode " + inode.getId() + " to not be write locked",
             mInodeLockManager.inodeWriteLockedByCurrentThread(inode.getId()));
@@ -90,14 +90,14 @@ public class BaseInodeLockingTest {
   /**
    * Checks that only the edges leading to the specified inodes are read-locked.
    */
-  protected void checkOnlyIncomingEdgesReadLocked(ReadOnlyInode... inodes) {
-    HashSet<ReadOnlyInode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
-    for (ReadOnlyInode inode : inodes) {
+  protected void checkOnlyIncomingEdgesReadLocked(Inode... inodes) {
+    HashSet<Inode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
+    for (Inode inode : inodes) {
       Edge edge = new Edge(inode.getParentId(), inode.getName());
       assertTrue("Expected edge " + edge + " to be read locked",
           mInodeLockManager.edgeReadLockedByCurrentThread(edge));
     }
-    for (ReadOnlyInode inode : mAllInodes) {
+    for (Inode inode : mAllInodes) {
       if (!shouldBeLocked.contains(inode)) {
         Edge edge = new Edge(inode.getParentId(), inode.getName());
         assertFalse("Expected edge " + edge + " to not be read locked",
@@ -109,14 +109,14 @@ public class BaseInodeLockingTest {
   /**
    * Checks that only the edges leading to the specified inodes are write-locked.
    */
-  protected void checkOnlyIncomingEdgesWriteLocked(ReadOnlyInode... inodes) {
-    HashSet<ReadOnlyInode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
-    for (ReadOnlyInode inode : inodes) {
+  protected void checkOnlyIncomingEdgesWriteLocked(Inode... inodes) {
+    HashSet<Inode> shouldBeLocked = new HashSet<>(Arrays.asList(inodes));
+    for (Inode inode : inodes) {
       Edge edge = new Edge(inode.getParentId(), inode.getName());
       assertTrue("Expected edge " + edge + " to be write locked",
           mInodeLockManager.edgeWriteLockedByCurrentThread(edge));
     }
-    for (ReadOnlyInode inode : mAllInodes) {
+    for (Inode inode : mAllInodes) {
       if (!shouldBeLocked.contains(inode)) {
         Edge edge = new Edge(inode.getParentId(), inode.getName());
         assertFalse("Expected edge " + edge + " to not be write locked",
@@ -134,20 +134,20 @@ public class BaseInodeLockingTest {
         mInodeLockManager.edgeWriteLockedByCurrentThread(edge));
   }
 
-  protected ReadOnlyInodeDirectory inodeDir(long id, long parentId, String name,
-      ReadOnlyInode... children) {
-    InodeDirectory dir =
-        InodeDirectory.create(id, parentId, name, CreateDirectoryOptions.defaults());
+  protected InodeDirectory inodeDir(long id, long parentId, String name, Inode... children) {
+    MutableInodeDirectory dir =
+        MutableInodeDirectory.create(id, parentId, name, CreateDirectoryOptions.defaults());
     mInodeStore.writeInode(dir);
-    for (ReadOnlyInode child : children) {
+    for (Inode child : children) {
       mInodeStore.addChild(dir.getId(), child);
     }
-    return ReadOnlyInode.wrap(dir).asDirectory();
+    return Inode.wrap(dir).asDirectory();
   }
 
-  protected ReadOnlyInodeFile inodeFile(long id, long parentId, String name) {
-    InodeFile file = InodeFile.create(id, parentId, name, 0, CreateFileOptions.defaults());
+  protected InodeFile inodeFile(long id, long parentId, String name) {
+    MutableInodeFile file =
+        MutableInodeFile.create(id, parentId, name, 0, CreateFileOptions.defaults());
     mInodeStore.writeInode(file);
-    return ReadOnlyInode.wrap(file).asFile();
+    return Inode.wrap(file).asFile();
   }
 }
