+++ core/server/master/src/test/java/alluxio/master/file/meta/MutableInodeDirectoryTest.java
@@ -22,26 +22,29 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Unit tests for {@link InodeDirectory}.
+ * Unit tests for {@link MutableInodeDirectory}.
  */
-public final class InodeDirectoryTest extends AbstractInodeTest {
-  private static final Logger LOG = LoggerFactory.getLogger(InodeDirectoryTest.class);
+public final class MutableInodeDirectoryTest extends AbstractInodeTest {
+  private static final Logger LOG = LoggerFactory.getLogger(MutableInodeDirectoryTest.class);
 
   /**
-   * Tests the {@link InodeDirectory#equals(Object)} method.
+   * Tests the {@link MutableInodeDirectory#equals(Object)} method.
    */
   @Test
   public void equalsTest() throws Exception {
-    InodeDirectory inode1 = InodeDirectory.create(1, 0, "test1", CreateDirectoryOptions.defaults());
-    InodeDirectory inode2 = InodeDirectory.create(1, 0, "test2", CreateDirectoryOptions.defaults());
-    InodeDirectory inode3 = InodeDirectory.create(3, 0, "test3", CreateDirectoryOptions.defaults());
+    MutableInodeDirectory inode1 =
+        MutableInodeDirectory.create(1, 0, "test1", CreateDirectoryOptions.defaults());
+    MutableInodeDirectory inode2 =
+        MutableInodeDirectory.create(1, 0, "test2", CreateDirectoryOptions.defaults());
+    MutableInodeDirectory inode3 =
+        MutableInodeDirectory.create(3, 0, "test3", CreateDirectoryOptions.defaults());
     Assert.assertTrue(inode1.equals(inode2));
     Assert.assertTrue(inode1.equals(inode1));
     Assert.assertFalse(inode1.equals(inode3));
   }
 
   /**
-   * Tests the {@link InodeDirectory#getId()} method.
+   * Tests the {@link MutableInodeDirectory#getId()} method.
    */
   @Test
   public void getId() {
@@ -49,7 +52,7 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#isDirectory()} method.
+   * Tests the {@link MutableInodeDirectory#isDirectory()} method.
    */
   @Test
   public void isDirectory() {
@@ -57,7 +60,7 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#isFile()} method.
+   * Tests the {@link MutableInodeDirectory#isFile()} method.
    */
   @Test
   public void isFile() {
@@ -65,11 +68,11 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#setDeleted(boolean)} method.
+   * Tests the {@link MutableInodeDirectory#setDeleted(boolean)} method.
    */
   @Test
   public void deleteInode() {
-    InodeDirectory inode1 = createInodeDirectory();
+    MutableInodeDirectory inode1 = createInodeDirectory();
     Assert.assertFalse(inode1.isDeleted());
     inode1.setDeleted(true);
     Assert.assertTrue(inode1.isDeleted());
@@ -83,7 +86,7 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   @Test
   public void initialLastModificationTime() {
     long lowerBoundMs = System.currentTimeMillis();
-    InodeDirectory inodeDirectory = createInodeDirectory();
+    MutableInodeDirectory inodeDirectory = createInodeDirectory();
     long upperBoundMs = System.currentTimeMillis();
     long lastModificationTimeMs = inodeDirectory.getLastModificationTimeMs();
     Assert.assertTrue(lowerBoundMs <= lastModificationTimeMs);
@@ -91,11 +94,11 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#setLastModificationTimeMs(long)} method.
+   * Tests the {@link MutableInodeDirectory#setLastModificationTimeMs(long)} method.
    */
   @Test
   public void setLastModificationTime() {
-    InodeDirectory inodeDirectory = createInodeDirectory();
+    MutableInodeDirectory inodeDirectory = createInodeDirectory();
     long lastModificationTimeMs = inodeDirectory.getLastModificationTimeMs();
     long newLastModificationTimeMs = lastModificationTimeMs + Constants.SECOND_MS;
     inodeDirectory.setLastModificationTimeMs(newLastModificationTimeMs);
@@ -103,12 +106,12 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#setLastModificationTimeMs(long)} method when setting an
+   * Tests the {@link MutableInodeDirectory#setLastModificationTimeMs(long)} method when setting an
    * invalid time.
    */
   @Test
   public void setInvalidLastModificationTime() {
-    InodeDirectory inodeDirectory = createInodeDirectory();
+    MutableInodeDirectory inodeDirectory = createInodeDirectory();
     long lastModificationTimeMs = inodeDirectory.getLastModificationTimeMs();
     long invalidModificationTimeMs = lastModificationTimeMs - Constants.SECOND_MS;
     inodeDirectory.setLastModificationTimeMs(invalidModificationTimeMs);
@@ -116,44 +119,44 @@ public final class InodeDirectoryTest extends AbstractInodeTest {
   }
 
   /**
-   * Tests the {@link InodeDirectory#setName(String)} method.
+   * Tests the {@link MutableInodeDirectory#setName(String)} method.
    */
   @Test
   public void setName() {
-    InodeDirectory inode1 = createInodeDirectory();
+    MutableInodeDirectory inode1 = createInodeDirectory();
     Assert.assertEquals("test1", inode1.getName());
     inode1.setName("test2");
     Assert.assertEquals("test2", inode1.getName());
   }
 
   /**
-   * Tests the {@link InodeDirectory#setParentId(long)} method.
+   * Tests the {@link MutableInodeDirectory#setParentId(long)} method.
    */
   @Test
   public void setParentId() {
-    InodeDirectory inode1 = createInodeDirectory();
+    MutableInodeDirectory inode1 = createInodeDirectory();
     Assert.assertEquals(0, inode1.getParentId());
     inode1.setParentId(2);
     Assert.assertEquals(2, inode1.getParentId());
   }
 
   /**
-   * Tests the {@link InodeDirectory#getMode()} method.
+   * Tests the {@link MutableInodeDirectory#getMode()} method.
    */
   @Test
   public void permissionStatus() {
-    InodeDirectory inode2 = createInodeDirectory();
+    MutableInodeDirectory inode2 = createInodeDirectory();
     Assert.assertEquals(TEST_OWNER, inode2.getOwner());
     Assert.assertEquals(TEST_GROUP, inode2.getGroup());
     Assert.assertEquals(Mode.defaults().applyDirectoryUMask().toShort(), inode2.getMode());
   }
 
   /**
-   * Tests the {@link InodeDirectory#generateClientFileInfo(String)} method.
+   * Tests the {@link MutableInodeDirectory#generateClientFileInfo(String)} method.
    */
   @Test
   public void generateClientFileInfo() {
-    InodeDirectory inodeDirectory = createInodeDirectory();
+    MutableInodeDirectory inodeDirectory = createInodeDirectory();
     String path = "/test/path";
     FileInfo info = inodeDirectory.generateClientFileInfo(path);
     Assert.assertEquals(inodeDirectory.getId(), info.getFileId());
